(all) Rename project and namespace
- `SiteGenerator` is overly long and complex. The project is de facto
  called "sitegen", so let's just go wit that for now (in PascalCase).

- Likewise, `SiteGenerator.ConsoleApp` is a bit overly lengthy. Let's
  remove the `ConsoleApp` suffix. This will play well with the
  `Sitegen.Tests` project coming up shortly.
