--- SiteGenerator.ConsoleApp/Models/LineBreaks.cs
@@ -1,8 +1,0 @@
-namespace SiteGenerator.ConsoleApp.Models
-{
-    public enum LineBreaks
-    {
-        Soft,
-        Hard
-    }
-}
