+++ src/Sitegen/Models/LineBreaks.cs
@@ -1,0 +1,13 @@
+using JetBrains.Annotations;
+
+namespace Sitegen.Models
+{
+    public enum LineBreaks
+    {
+        [UsedImplicitly]
+        Soft,
+
+        [UsedImplicitly]
+        Hard
+    }
+}
