+++ src/Sitegen/Sitegen.csproj
@@ -4,6 +4,7 @@
         <OutputType>Exe</OutputType>
         <TargetFramework>net5.0</TargetFramework>
         <AssemblyName>sitegen</AssemblyName>
+        <RootNamespace>Sitegen</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
